+++ nuxeo-drive-client/nxdrive/engine/next/queue_manager.py
@@ -15,10 +15,7 @@ class QueueManager(OldQueueManager):
     def postpone_pair(self, doc_pair, interval=60):
         doc_pair.error_next_try = interval + int(time.time())
         log.debug("Blacklisting pair for %ds: %r", interval, doc_pair)
-        self._error_lock.acquire()
-        try:
+        with self._error_lock:
             self._on_error_queue[doc_pair.id] = doc_pair
             if not self._error_timer.isActive():
                 self.newError.emit(doc_pair.id)
-        finally:
-            self._error_lock.release()